功能优化完善，修复bug

+++ src/com/interactive/classroom/dao/impl/AttendanceDaoImpl.java
@@ -3,6 +3,7 @@ package com.interactive.classroom.dao.impl;
 import com.interactive.classroom.bean.AttendanceBean;
 import com.interactive.classroom.constant.UserType;
 import com.interactive.classroom.dao.AttendanceDao;
+import com.interactive.classroom.dao.UserDao;
 import com.interactive.classroom.dao.filters.AttendanceFilter;
 import com.interactive.classroom.utils.DatabaseHelper;
 import com.interactive.classroom.utils.Log;
@@ -78,6 +79,47 @@ public class AttendanceDaoImpl implements AttendanceDao {
     }
 
     @Override
+    public JSONObject queryAttendanceUsers(String courseId, String attendanceFlag) throws SQLException, JSONException {
+        String sql = "select user_manage.*,course_student.*" +
+                " from user_manage,course_student" +
+                " where course_student.course_id=" + courseId +
+                " and user_manage.id=course_student.student_id" +
+                " group by user_manage.id";
+        ResultSet rs = DatabaseHelper.executeQuery(sql);
+        JSONArray jsonArray = new JSONArray();
+        while (rs.next()) {
+            JSONObject jsonObject = new JSONObject();
+            String name = rs.getString("name");
+            String grade = rs.getString("grade");
+            String className = rs.getString("class");
+            String faculty = rs.getString("faculty");
+            String studentNum = rs.getString("student_num");
+            jsonObject.put("name", name == null ? "未知" : name);
+            jsonObject.put("grade", grade == null ? "未知" : grade);
+            jsonObject.put("class", className == null ? "未知" : className);
+            jsonObject.put("faculty", faculty == null ? "未知" : faculty);
+            jsonObject.put("student_num", studentNum == null ? "未知" : studentNum);
+            String attendance = rs.getString("attendance");
+            boolean hasAttendanced = false;
+            if (attendance != null) {
+                for (String flag : attendance.split(",")) {
+                    if (attendanceFlag.equals(flag)) {
+                        hasAttendanced = true;
+                        break;
+                    }
+                }
+            }
+            jsonObject.put("has_attendanced", hasAttendanced);
+            jsonArray.put(jsonObject);
+        }
+        JSONObject jsonObject = new JSONObject();
+        jsonObject.put("aaData", jsonArray);
+        jsonObject.put("result_msg", "ok");
+        jsonObject.put("result_code", 0);
+        return jsonObject;
+    }
+
+    @Override
     public JSONObject updateAttendanceInfo(AttendanceBean attendance) throws JSONException {
         String sql = "UPDATE " + TABLE_NAME + " set publisher_name=?,attendance_title=?,attendance_requirement=?,deadline=? where id=" + attendance.getId();
         boolean isSuccess = DatabaseHelper.executeUpdate(
@@ -102,24 +144,41 @@ public class AttendanceDaoImpl implements AttendanceDao {
                 + bean.getCourseId() + ","
                 + bean.getAttendanceFlag() + ")";
         Log.d(getClass().getName(), "publishAttendance sql=" + sql);
-        DatabaseHelper.executeUpdate(sql);
+        boolean isSuccess = DatabaseHelper.executeUpdate(sql);
+        if (isSuccess) {
+            sql = "UPDATE course_manage set attendance_count=attendance_count+1 where id=" + bean.getCourseId();
+            isSuccess = DatabaseHelper.executeUpdate(sql);
+        }
         JSONObject jsonObj = new JSONObject();
-        jsonObj.put("result_msg", "ok");
-        jsonObj.put("result_code", 0);
+        jsonObj.put("result_msg", isSuccess ? "ok" : "发布失败！");
+        jsonObj.put("result_code", isSuccess ? 0 : 10);
         return jsonObj;
     }
 
     @Override
-    public JSONObject deleteAttendance(String[] ids) throws JSONException {
-        String resultMsg = "ok";
-        int resultCode = 0;
+    public JSONObject deleteAttendance(String[] ids) throws JSONException, SQLException {
+        boolean isSuccess = false;
         for (String id : ids) {
-            String sql = "delete from " + TABLE_NAME + " where id=" + id;
-            DatabaseHelper.executeUpdate(sql);
+            String sql = "select * from " + TABLE_NAME + " where id=" + id;
+            ResultSet rs = DatabaseHelper.executeQuery(sql);
+            String courseId;
+            if (rs.next()) {
+                courseId = rs.getString("course_id");
+                int attendanceFlag = rs.getInt("attendance_flag");
+                Log.d(getClass().getName(), "course_id=" + courseId);
+                sql = "delete from " + TABLE_NAME + " where id=" + id;
+                isSuccess = DatabaseHelper.executeUpdate(sql);
+                if (isSuccess) {
+                    sql = "UPDATE course_manage set attendance_count=attendance_count-1 where id=" + courseId;
+                    DatabaseHelper.executeUpdate(sql);
+                    sql = "update course_student set attendance=replace(attendance,'" + attendanceFlag + "','') where course_id=" + courseId;
+                    isSuccess = DatabaseHelper.executeUpdate(sql);
+                }
+            }
         }
         JSONObject jsonObj = new JSONObject();
-        jsonObj.put("result_msg", resultMsg);
-        jsonObj.put("result_code", resultCode);
+        jsonObj.put("result_msg", isSuccess ? "ok" : "出错了！");
+        jsonObj.put("result_code", isSuccess ? 0 : 10);
         return jsonObj;
     }
